Refactor delete label effect
